requirements.txt
# Core Dependencies
streamlit
streamlit-node-editor
loguru
torch
GitPython

# Development & QA
mypy
ruff
pytest
sphinx
sphinx-rtd-theme
